New Line Scan Algorithm 1
This is a line scan algorithm that is an attempt to make the line search
more stable. This is currently untested, but the basic algorithm is to
search for a flat slope after detecting a maximum or minimum. We still
need an algorithm to figure out what happens when the line is at the
edge of the camera's vision.